# src/generated/openapi/java/com/example/openapi/api/CarsApi.java
/*
 * Persons API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.openapi.api;

import io.micronaut.http.annotation.*;
import io.micronaut.core.annotation.Nullable;
import io.micronaut.core.convert.format.Format;
import com.example.openapi.model.Car;
import com.example.openapi.model.CarCreateRequest;
import com.example.openapi.model.CarOwnerAssignRequest;
import com.example.openapi.model.CarSearchRequest;
import com.example.openapi.model.CarUpdateRequest;
import com.example.openapi.model.CarsResponse;
import io.micronaut.http.HttpResponse;
import static io.micronaut.core.convert.converters.MultiValuesConverterFactory.*;
import jakarta.annotation.Generated;
import java.util.ArrayList;
import java.util.HashMap;
import java.util.List;
import java.util.Map;
import jakarta.validation.Valid;
import jakarta.validation.constraints.*;

@Generated("io.micronaut.openapi.generator.JavaMicronautServerCodegen")
@Controller
public interface CarsApi {

    /**
     * assignCarOwner
     *
     * @param id (required)
     * @param carOwnerAssignRequest (optional)
     *
     * @return Car with owner assigned (status code 200)
     */
    @Put("/cars/{id}/owner")
    HttpResponse<@Valid Car> assignCarOwner(
        @PathVariable("id") @NotNull Integer id,
        @Body @Nullable(inherited = true) @Valid CarOwnerAssignRequest carOwnerAssignRequest
    );

    /**
     * createCar
     *
     * @param carCreateRequest (optional)
     *
     * @return Created car (with its issuer firm) (status code 200)
     */
    @Post("/cars")
    HttpResponse<@Valid Car> createCar(
        @Body @Nullable(inherited = true) @Valid CarCreateRequest carCreateRequest
    );

    /**
     * deleteCar
     *
     * @param id (required)
     *
     * @return Deleted (status code 204)
     */
    @Delete("/cars/{id}")
    HttpResponse<Void> deleteCar(
        @PathVariable("id") @NotNull Integer id
    );

    /**
     * getCar
     *
     * @param id (required)
     *
     * @return A car (status code 200)
     */
    @Get("/cars/{id}")
    HttpResponse<@Valid Car> getCar(
        @PathVariable("id") @NotNull Integer id
    );

    /**
     * searchCars
     *
     * @param page (optional)
     * @param size (optional)
     * @param sort (optional)
     * @param carSearchRequest (optional)
     *
     * @return Page of cars (status code 200)
     */
    @Post("/cars/search")
    HttpResponse<@Valid CarsResponse> searchCars(
        @QueryValue("page") @Nullable(inherited = true) Integer page,
        @QueryValue("size") @Nullable(inherited = true) Integer size,
        @QueryValue("sort") @Nullable(inherited = true) @Format(FORMAT_MULTI) List<@NotNull String> sort,
        @Body @Nullable(inherited = true) @Valid CarSearchRequest carSearchRequest
    );

    /**
     * updateCar
     *
     * @param id (required)
     * @param carUpdateRequest (optional)
     *
     * @return Updated car (status code 200)
     */
    @Put("/cars/{id}")
    HttpResponse<@Valid Car> updateCar(
        @PathVariable("id") @NotNull Integer id,
        @Body @Nullable(inherited = true) @Valid CarUpdateRequest carUpdateRequest
    );

}