# src/generated/openapi/java/com/example/openapi/api/InsurancesApi.java
/*
 * Persons API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.openapi.api;

import io.micronaut.http.annotation.*;
import io.micronaut.core.annotation.Nullable;
import io.micronaut.core.convert.format.Format;
import io.micronaut.http.HttpResponse;
import com.example.openapi.model.Insurance;
import com.example.openapi.model.InsuranceCreateRequest;
import com.example.openapi.model.InsuranceSearchRequest;
import com.example.openapi.model.InsuranceUpdateRequest;
import com.example.openapi.model.InsurancesResponse;
import static io.micronaut.core.convert.converters.MultiValuesConverterFactory.*;
import jakarta.annotation.Generated;
import java.util.ArrayList;
import java.util.HashMap;
import java.util.List;
import java.util.Map;
import jakarta.validation.Valid;
import jakarta.validation.constraints.*;

@Generated("io.micronaut.openapi.generator.JavaMicronautServerCodegen")
@Controller
public interface InsurancesApi {

    /**
     * createInsurance
     *
     * @param insuranceCreateRequest (optional)
     *
     * @return Created insurance (status code 200)
     */
    @Post("/insurances")
    HttpResponse<@Valid Insurance> createInsurance(
        @Body @Nullable(inherited = true) @Valid InsuranceCreateRequest insuranceCreateRequest
    );

    /**
     * deleteInsurance
     *
     * @param id (required)
     *
     * @return Deleted (status code 204)
     */
    @Delete("/insurances/{id}")
    HttpResponse<Void> deleteInsurance(
        @PathVariable("id") @NotNull Integer id
    );

    /**
     * getExpiringSoonInsurances
     *
     * @return Insurances with fewer than 30 days left (status code 200)
     */
    @Get("/insurances/expiring-soon")
    HttpResponse<@NotNull List<@Valid Insurance>> getExpiringSoonInsurances();

    /**
     * getInsurance
     *
     * @param id (required)
     *
     * @return An insurance (status code 200)
     */
    @Get("/insurances/{id}")
    HttpResponse<@Valid Insurance> getInsurance(
        @PathVariable("id") @NotNull Integer id
    );

    /**
     * searchInsurances
     *
     * @param page (optional)
     * @param size (optional)
     * @param sort (optional)
     * @param insuranceSearchRequest (optional)
     *
     * @return Page of insurances (status code 200)
     */
    @Post("/insurances/search")
    HttpResponse<@Valid InsurancesResponse> searchInsurances(
        @QueryValue("page") @Nullable(inherited = true) Integer page,
        @QueryValue("size") @Nullable(inherited = true) Integer size,
        @QueryValue("sort") @Nullable(inherited = true) @Format(FORMAT_MULTI) List<@NotNull String> sort,
        @Body @Nullable(inherited = true) @Valid InsuranceSearchRequest insuranceSearchRequest
    );

    /**
     * updateInsurance
     *
     * @param id (required)
     * @param insuranceUpdateRequest (optional)
     *
     * @return Updated insurance (status code 200)
     */
    @Put("/insurances/{id}")
    HttpResponse<@Valid Insurance> updateInsurance(
        @PathVariable("id") @NotNull Integer id,
        @Body @Nullable(inherited = true) @Valid InsuranceUpdateRequest insuranceUpdateRequest
    );

}