# src/generated/openapi/java/com/example/openapi/api/IssuerFirmsApi.java
/*
 * Persons API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.openapi.api;

import io.micronaut.http.annotation.*;
import io.micronaut.core.annotation.Nullable;
import io.micronaut.core.convert.format.Format;
import io.micronaut.http.HttpResponse;
import com.example.openapi.model.IssuerFirmSearchRequest;
import com.example.openapi.model.IssuerFirmsResponse;
import static io.micronaut.core.convert.converters.MultiValuesConverterFactory.*;
import jakarta.annotation.Generated;
import java.util.ArrayList;
import java.util.HashMap;
import java.util.List;
import java.util.Map;
import jakarta.validation.Valid;
import jakarta.validation.constraints.*;

@Generated("io.micronaut.openapi.generator.JavaMicronautServerCodegen")
@Controller
public interface IssuerFirmsApi {

    /**
     * searchIssuerFirms
     *
     * @param page (optional)
     * @param size (optional)
     * @param sort (optional)
     * @param issuerFirmSearchRequest (optional)
     *
     * @return Page of issuer firms (status code 200)
     */
    @Post("/issuer-firms/search")
    HttpResponse<@Valid IssuerFirmsResponse> searchIssuerFirms(
        @QueryValue("page") @Nullable(inherited = true) Integer page,
        @QueryValue("size") @Nullable(inherited = true) Integer size,
        @QueryValue("sort") @Nullable(inherited = true) @Format(FORMAT_MULTI) List<@NotNull String> sort,
        @Body @Nullable(inherited = true) @Valid IssuerFirmSearchRequest issuerFirmSearchRequest
    );

}