# src/generated/openapi/java/com/example/openapi/api/PersonsApi.java
/*
 * Persons API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.openapi.api;

import io.micronaut.http.annotation.*;
import io.micronaut.core.annotation.Nullable;
import io.micronaut.core.convert.format.Format;
import io.micronaut.http.HttpResponse;
import com.example.openapi.model.Person;
import com.example.openapi.model.PersonCreateRequest;
import com.example.openapi.model.PersonSearchRequest;
import com.example.openapi.model.PersonsResponse;
import static io.micronaut.core.convert.converters.MultiValuesConverterFactory.*;
import jakarta.annotation.Generated;
import java.util.ArrayList;
import java.util.HashMap;
import java.util.List;
import java.util.Map;
import jakarta.validation.Valid;
import jakarta.validation.constraints.*;

@Generated("io.micronaut.openapi.generator.JavaMicronautServerCodegen")
@Controller
public interface PersonsApi {

    /**
     * createPerson
     *
     * @param personCreateRequest (optional)
     *
     * @return Created person (status code 200)
     */
    @Post("/persons")
    HttpResponse<@Valid Person> createPerson(
        @Body @Nullable(inherited = true) @Valid PersonCreateRequest personCreateRequest
    );

    /**
     * getPerson
     *
     * @param id (required)
     *
     * @return A person (status code 200)
     */
    @Get("/persons/{id}")
    HttpResponse<@Valid Person> getPerson(
        @PathVariable("id") @NotNull Integer id
    );

    /**
     * getPersons
     *
     * @return A person (status code 200)
     */
    @Get("/persons")
    HttpResponse<@NotNull List<@Valid Person>> getPersons();

    /**
     * searchPersons
     *
     * @param page (optional)
     * @param size (optional)
     * @param sort (optional)
     * @param personSearchRequest (optional)
     *
     * @return Page of persons (status code 200)
     */
    @Post("/persons/search")
    HttpResponse<@Valid PersonsResponse> searchPersons(
        @QueryValue("page") @Nullable(inherited = true) Integer page,
        @QueryValue("size") @Nullable(inherited = true) Integer size,
        @QueryValue("sort") @Nullable(inherited = true) @Format(FORMAT_MULTI) List<@NotNull String> sort,
        @Body @Nullable(inherited = true) @Valid PersonSearchRequest personSearchRequest
    );

    /**
     * updatePerson
     *
     * @param id (required)
     * @param person (optional)
     *
     * @return Updated (status code 200)
     */
    @Put("/persons/{id}")
    HttpResponse<Void> updatePerson(
        @PathVariable("id") @NotNull Integer id,
        @Body @Nullable(inherited = true) @Valid Person person
    );

}