# src/generated/openapi/java/com/example/openapi/model/Car.java
/*
 * Persons API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.openapi.model;

import java.util.Objects;
import java.util.Arrays;
import java.util.Map;
import java.util.function.Function;
import java.util.stream.Collectors;
import com.fasterxml.jackson.annotation.*;
import io.micronaut.serde.annotation.Serdeable;
import jakarta.validation.constraints.*;
import jakarta.validation.Valid;
import io.micronaut.core.annotation.Nullable;
import jakarta.annotation.Generated;

/**
 * Car
 */
@JsonPropertyOrder({
    Car.JSON_PROPERTY_ID,
    Car.JSON_PROPERTY_MARK,
    Car.JSON_PROPERTY_MODEL,
    Car.JSON_PROPERTY_OWNER_ID,
    Car.JSON_PROPERTY_OWNER_NAME,
    Car.JSON_PROPERTY_ISSUER_FIRM_NAME,
})
@Serdeable
@Generated("io.micronaut.openapi.generator.JavaMicronautServerCodegen")
public class Car {

    public static final String JSON_PROPERTY_ID = "id";
    public static final String JSON_PROPERTY_MARK = "mark";
    public static final String JSON_PROPERTY_MODEL = "model";
    public static final String JSON_PROPERTY_OWNER_ID = "ownerId";
    public static final String JSON_PROPERTY_OWNER_NAME = "ownerName";
    public static final String JSON_PROPERTY_ISSUER_FIRM_NAME = "issuerFirmName";

    @NotNull
    @JsonProperty(JSON_PROPERTY_ID)
    private Integer id;

    @NotNull
    @JsonProperty(JSON_PROPERTY_MARK)
    private String mark;

    @NotNull
    @JsonProperty(JSON_PROPERTY_MODEL)
    private String model;

    @Nullable(inherited = true)
    @JsonProperty(JSON_PROPERTY_OWNER_ID)
    @JsonInclude(JsonInclude.Include.USE_DEFAULTS)
    private Integer ownerId;

    @Nullable(inherited = true)
    @JsonProperty(JSON_PROPERTY_OWNER_NAME)
    @JsonInclude(JsonInclude.Include.USE_DEFAULTS)
    private String ownerName;

    @Nullable(inherited = true)
    @JsonProperty(JSON_PROPERTY_ISSUER_FIRM_NAME)
    @JsonInclude(JsonInclude.Include.USE_DEFAULTS)
    private String issuerFirmName;

    public Car(Integer id, String mark, String model) {
        this.id = id;
        this.mark = mark;
        this.model = model;
    }

    /**
     * @return the id property value
     */
    public Integer getId() {
        return id;
    }

    /**
     * Set the id property value
     *
     * @param id property value to set
     */
    public void setId(Integer id) {
        this.id = id;
    }

    /**
     * Set id in a chainable fashion.
     *
     * @return The same instance of Car for chaining.
     */
    public Car id(Integer id) {
        this.id = id;
        return this;
    }

    /**
     * @return the mark property value
     */
    public String getMark() {
        return mark;
    }

    /**
     * Set the mark property value
     *
     * @param mark property value to set
     */
    public void setMark(String mark) {
        this.mark = mark;
    }

    /**
     * Set mark in a chainable fashion.
     *
     * @return The same instance of Car for chaining.
     */
    public Car mark(String mark) {
        this.mark = mark;
        return this;
    }

    /**
     * @return the model property value
     */
    public String getModel() {
        return model;
    }

    /**
     * Set the model property value
     *
     * @param model property value to set
     */
    public void setModel(String model) {
        this.model = model;
    }

    /**
     * Set model in a chainable fashion.
     *
     * @return The same instance of Car for chaining.
     */
    public Car model(String model) {
        this.model = model;
        return this;
    }

    /**
     * @return the ownerId property value
     */
    public Integer getOwnerId() {
        return ownerId;
    }

    /**
     * Set the ownerId property value
     *
     * @param ownerId property value to set
     */
    public void setOwnerId(Integer ownerId) {
        this.ownerId = ownerId;
    }

    /**
     * Set ownerId in a chainable fashion.
     *
     * @return The same instance of Car for chaining.
     */
    public Car ownerId(Integer ownerId) {
        this.ownerId = ownerId;
        return this;
    }

    /**
     * @return the ownerName property value
     */
    public String getOwnerName() {
        return ownerName;
    }

    /**
     * Set the ownerName property value
     *
     * @param ownerName property value to set
     */
    public void setOwnerName(String ownerName) {
        this.ownerName = ownerName;
    }

    /**
     * Set ownerName in a chainable fashion.
     *
     * @return The same instance of Car for chaining.
     */
    public Car ownerName(String ownerName) {
        this.ownerName = ownerName;
        return this;
    }

    /**
     * @return the issuerFirmName property value
     */
    public String getIssuerFirmName() {
        return issuerFirmName;
    }

    /**
     * Set the issuerFirmName property value
     *
     * @param issuerFirmName property value to set
     */
    public void setIssuerFirmName(String issuerFirmName) {
        this.issuerFirmName = issuerFirmName;
    }

    /**
     * Set issuerFirmName in a chainable fashion.
     *
     * @return The same instance of Car for chaining.
     */
    public Car issuerFirmName(String issuerFirmName) {
        this.issuerFirmName = issuerFirmName;
        return this;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) {
            return true;
        }
        if (o == null || getClass() != o.getClass()) {
            return false;
        }
        var car = (Car) o;
        return Objects.equals(this.id, car.id)
            && Objects.equals(this.mark, car.mark)
            && Objects.equals(this.model, car.model)
            && Objects.equals(this.ownerId, car.ownerId)
            && Objects.equals(this.ownerName, car.ownerName)
            && Objects.equals(this.issuerFirmName, car.issuerFirmName);
    }

    @Override
    public int hashCode() {
        return Objects.hash(id, mark, model, ownerId, ownerName, issuerFirmName);
    }

    @Override
    public String toString() {
        return "Car("
            + "id: " + getId() + ", "
            + "mark: " + getMark() + ", "
            + "model: " + getModel() + ", "
            + "ownerId: " + getOwnerId() + ", "
            + "ownerName: " + getOwnerName() + ", "
            + "issuerFirmName: " + getIssuerFirmName()
            + ")";
    }

}